Isolate Regexp functions into its own runtime module
