+++ opal/runtime/boot.js
@@ -2088,89 +2088,6 @@
     }
   };
 
-  // Regexps
-  // -------
-
-  // Escape Regexp special chars letting the resulting string be used to build
-  // a new Regexp.
-  //
-  Opal.escape_regexp = function(str) {
-    return Opal.escape_metacharacters(str.replace(/([-[\]\/{}()*+?.^$\\| ])/g, '\\$1'))
-  };
-
-  Opal.escape_metacharacters = function(str) {
-    return str.replace(/[\n]/g, '\\n')
-              .replace(/[\r]/g, '\\r')
-              .replace(/[\f]/g, '\\f')
-              .replace(/[\t]/g, '\\t');
-  }
-
-  // Create a global Regexp from a RegExp object and cache the result
-  // on the object itself ($$g attribute).
-  //
-  Opal.global_regexp = function(pattern) {
-    if (pattern.global) {
-      return pattern; // RegExp already has the global flag
-    }
-    if (pattern.$$g == null) {
-      pattern.$$g = new RegExp(pattern.source, 'g' + pattern.flags);
-    } else {
-      pattern.$$g.lastIndex = null; // reset lastIndex property
-    }
-    return pattern.$$g;
-  };
-
-  // Transform a regular expression from Ruby syntax to JS syntax.
-  Opal.transform_regexp = function(regexp, flags) {
-    return Opal.Opal.RegexpTranspiler.$transform_regexp(regexp, flags);
-  };
-
-  // Combine multiple regexp parts together
-  Opal.regexp = function(parts, flags) {
-    var part;
-
-    if (flags == null) flags = '';
-
-    var ignoreCase = flags.includes('i');
-
-    for (var i = 0, ii = parts.length; i < ii; i++) {
-      part = parts[i];
-      if (part instanceof RegExp) {
-        if (part.ignoreCase !== ignoreCase)
-          Opal.Kernel.$warn(
-            "ignore case doesn't match for " + part.source.$inspect(),
-            new Map([['uplevel',  1]])
-          )
-
-        part = part.$$source != null ? part.$$source : part.source;
-      }
-      if (part == '') part = '(?:)';
-      parts[i] = part;
-    }
-
-    parts = parts.join('');
-    parts = Opal.escape_metacharacters(parts);
-
-    var output = Opal.transform_regexp(parts, flags);
-
-    var regexp = new RegExp(output[0], output[1]);
-    if (parts != output[0]) regexp.$$source = parts
-    if (flags != output[1]) regexp.$$options = flags;
-    return regexp;
-  };
-
-  // Regexp has been transformed, so let's annotate the original regexp
-  Opal.annotate_regexp = function(regexp, source, options) {
-    regexp.$$source = source;
-    regexp.$$options = options;
-    return regexp;
-  }
-
-  // Annotated empty regexp
-  Opal.empty_regexp = function(flags) {
-    return Opal.annotate_regexp(new RegExp(/(?:)/, flags), '');
-  }
-
   // Require system
   // --------------
