+++ opal/runtime/runtime.rb
@@ -2,4 +2,5 @@ require 'runtime/boot'
 require 'runtime/const'
 require 'runtime/send'
 require 'runtime/hash'
+require 'runtime/regexp'
 require 'runtime/helpers'
